# Conflicts:
#	src/core/CLucene/search/spans/NearSpansComplete.cpp

+++ src/core/CLucene/search/spans/NearSpansComplete.cpp
@@ -438,10 +438,10 @@ void NearSpansComplete::SpansCell::addEnds( int32_t currentEnd, int32_t minSlop,
     size_t sz = cache.size();
     for( size_t pos = 1; pos < sz; pos += 2 )
     {
-        end = cache.get( pos );
+        end = cache.get( (int32_t)pos );
         if( end > currentEnd )
         {
-            int32_t matchSlop = cache.get( pos-1 ) - slopBase;
+            int32_t matchSlop = cache.get( (int32_t)(pos-1) ) - slopBase;
             if( matchSlop <= maxSlop && minSlop <= matchSlop )
                 ends.insert( end );
         }
